+++ CHANGELOG.md
@@ -10,3 +10,5 @@
 * Pulled d3-tip into the source code to make modifications to add a way to determine the direction of the tip so that the tip doesn't go off the screen.
 * Minified the CSS stylesheet.
 * Added a way to configure the truncate cap in the configuration object.
+* Removed d3-tip from the dependencies.
+* Added AMD and CommonJS support.